python/pyproject.toml
[project]
name = "teleop_android"
version = "0.1.0"
description = "Start a WebSocket server for listening for the phone's position/orientation from an Android device."
requires-python = ">=3.13"
dependencies = ["fastapi>=0.121.0", "uvicorn[standard]>=0.38.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["teleop_android"]

[project.optional-dependencies]
rerun = ["rerun-sdk>=0.26.2", "numpy>=2.3.4", "scipy>=1.16.3"]
